+++ module-test/src/___foo/bar.rs
@@ -1,0 +1,10 @@
+//use crate::constants;
+//
+//pub fn test_bar() {
+//    println!("bar {}", constants::CONSTANT_A);
+//    test_bar_internal();
+//}
+//
+//fn test_bar_internal() {
+//    println!("test_bar_internal");
+//}
